Update styling, refactor screen size hook, and refactor scenario tester our of csv.
